+++ src/routes/routes.tsx
@@ -19,7 +19,8 @@ import PortfolioCreate from "../pages/mypage/specialist/portfolio/portfolioCreat
 // import TransactionIndex from "../pages/mypage/specialist/transaction/transactionIndex"
 // import Transaction from "../pages/mypage/specialist/transaction/transaction"
 import IssueCreate from "../pages/mypage/client/issue/createIssue"
-import IssueIndex from "../pages/mypage/client/issue/issue"
+import IssueIndex from "../pages/mypage/client/issue/issueIndex"
+import Issue from "../pages/mypage/client/issue/issue"
 
 const RootRouter: React.FC = () => {
     return (
@@ -60,6 +61,9 @@ const RootRouter: React.FC = () => {
                 </PrivateRoute>
                 <PrivateRoute path='/mypage/issue/index'>
                     <Mypage><IssueIndex /></Mypage>
+                </PrivateRoute>
+                <PrivateRoute path='/mypage/issue'>
+                    <Mypage><Issue /></Mypage>
                 </PrivateRoute>
                 <Route path='/mypage' component={Mypage} />
             </Switch>
